+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `free_from_str` and `free_from_fn`.
+
 ### Changed
 - `value_from_fn` allows any error type that implements `Display` now
   and not only `String`.
+- The `Error` enum.
 
 [Unreleased]: https://github.com/RazrFalcon/ttf-parser/compare/v0.1.0...HEAD